Make iI mapping do something useful
Prior to this commit the iI mapping was equivalent to ii. I thought this
was a waste so I added functionality which is equivalent to doing the
old aI mapping and then moving the top of the text object by one line.
Doing this makes it so we can select loops like this:

```
for(i = 1; i <= 10; i++)
{
	printf("%d\n", i);
}
```
I also restructured the mappings so that ii selects all lines of current
indent, iI selects ii and moves the top of the selection up by one, ai
does iI and moves the bottom of the selection down by one, and aI does
what I described.